fixed refactored imports in tests
