+++ test/suite/configuration.test.ts
@@ -1,7 +1,7 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
 import * as Constants from "../../src/Constants";
-import { ConfigurationManager } from "../../configuration/manager";
+import { ConfigurationManager } from "../../src/ConfigurationManager";
 
 suite("ConfigurationManager Test Suite", () => {
   test("ConfigurationManager should load configuration", () => {
