+++ test/suite/linter.test.ts
@@ -1,9 +1,9 @@
 import * as assert from "assert";
 import * as fs from "fs";
 import * as path from "path";
-import { ConfigurationManager } from "../../configuration/manager";
+import { ConfigurationManager } from "../../src/ConfigurationManager";
 import { IAnnotatedtext } from "annotatedtext";
-import { Linter } from "../../linter/linter";
+import { Linter } from "../../src/Linter";
 
 suite("Linter Test Suite", () => {
   const config: ConfigurationManager = new ConfigurationManager();
